Changed price in cart to int

+++ src/cart/src/main/java/works/weave/socks/cart/controllers/api/Item.java
@@ -9,7 +9,7 @@ public class Item {
 
     private int quantity;
 
-    private float unitPrice;
+    private int unitPrice;
 
     public static Item from(ItemEntity itemEntity) {
         Item item = new Item();
